refactor(projects): extract KitAddForm (Phase 4c)
Pull the inline "Add Kit to Project" dialog body out of equipment-tab.tsx
into kit-add-form.tsx so it can also be embedded inside the upcoming
UnifiedAddDialog (Phase 4d).

Pure refactor:
- equipment-tab.tsx kit dialog is now a 14-line wrapper that mounts
  <KitAddForm /> when open. Same trigger buttons (toolbar "Add Kit",
  group-kebab onAddKit) — no prop changes upstream.
- KitAddForm owns selectedKitId / pricingMode / unitPrice state, the
  kits + kit-availability queries, and the addKitLineItem mutation.
- Form receives projectId, rentalStartDate?, rentalEndDate?, categoryId?,
  groupId?, targetLabel?, onClose, onInvalidate — mirrors the
  EquipmentAddForm shape so Phase 4d can mount both behind one segmented
  switcher.
- Threaded rentalStartDate/rentalEndDate through to checkKitAvailability
  (previous inline implementation hardcoded `new Date(), new Date()`,
  ignoring project dates). Defaults to new Date() when undefined, so
  behavior is identical when dates aren't set yet.
- Removed dead state + queries + imports from equipment-tab.tsx
  (selectedKitId, kitPricingMode, kitUnitPrice, kitsData,
  kitAvailability, addKitMut, getKits, addKitLineItem,
  checkKitAvailability, ComboboxPicker).
- Verified: tsc clean, full unit suite passes, eslint shows only the
  pre-existing queryKey/useCallback warning, dialog opens with kit
  picker + Pricing Mode radio + conditional Unit Price input, mode
  toggle switches helper text correctly.
